+++ src/components/CopyButton/index.js
@@ -16,23 +16,32 @@ const CopyButton = ({
   testSection,
   textToCopy,
   usesTextLabel,
+  textLabel,
+  onCopy,
 }) => {
 
   return (
     <CopyToClipboard
-      text={ textToCopy }>
+      text={ textToCopy }
+      onCopy={ onCopy }>
       <Button
         style={ style === 'none' ? null : style }
         size={ usesTextLabel ? 'tiny' : null }
         ariaLabel="Copy code snippet"
         testSection={ testSection ? `${testSection}-copy-button` : null }>
-        { usesTextLabel ? 'Copy' : <Icon name='clipboard' /> }
+        { usesTextLabel ? textLabel : <Icon name='clipboard' /> }
       </Button>
     </CopyToClipboard>
   );
 };
 
 CopyButton.propTypes = {
+  /** Function to perform when content is copied.
+      This callback is passed as onCopy prop to CopyToClipboard component.
+      onCopy(text, result)
+      result (bool): Returns true if copied successfully, else false.
+  */
+  onCopy: PropTypes.func,
   /** Style option for the button */
   style: PropTypes.oneOf([
     'highlight',
@@ -48,10 +57,12 @@ CopyButton.propTypes = {
   ]),
   /** Hook for automated JavaScript tests */
   testSection: PropTypes.string,
+  /** String label to use when usesTextLabel is true */
+  textLabel: PropTypes.string,
   /** The text or code that will be copied */
   textToCopy: PropTypes.string.isRequired,
   /**
-   * Whether or not this should use the word copy
+   * Whether or not this should use a string label
    * instead of an icon
    */
   usesTextLabel: PropTypes.bool,
@@ -60,6 +71,7 @@ CopyButton.propTypes = {
 CopyButton.defaultProps = {
   style: 'plain',
   usesTextLabel: false,
+  textLabel: 'Copy',
 };
 
 export default CopyButton;